Move conversation file preview wrapper
